Remove the sgetspent method from the pwd impl.
FreeBSD and OS X do not implement an equivalent to sgetspent, and the
implementation included before did not work as expected.

sgetspent is supposed to parse a shadow user data in a string and return the first
entry. The implementation we had called getpwnam (in the pwd version), with the
supplied string which meant it was trying to find a user with a name equal to the
passed in shadow file string.

it's best to just remove the function all together.