Attempt to resolve BSD make issues.
The depend file was using ifneq statements in it, which was not
compatible with BSD's make.  Looking around at some other gems, it
doesn't seem to be standard to include ruby headers in the actual make
file.  This removes the check for ruby.h and ruby(/?)io.h in the
makefile, and adds a check into extconf.rb.