Fix up issue with shadow implementation.
When the implementations were split into shadow and pwd, the code
initialzing the structure used in the shadow implementation was replaced
by older code not including the added field, always nil in that
implemenation, for pw_change.

The intent currently is that the passwdEntry structure should contain
all the fields available across implementations.